requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-cloud-bigquery==3.26.0
google-cloud-firestore==2.16.0
firebase-admin==6.5.0
python-multipart==0.0.12
pydantic==2.9.0
email-validator==2.1.0
requests==2.31.0
